Renamed searcher to ranker
